refactor: board directory명 companion으로 변경
