--- src/main/java/com/project/Journey/board/paging/PostSpecification.java
@@ -1,45 +1,0 @@
-package com.project.Journey.board.paging;
-
-import com.project.Journey.board.dto.PostSearchRequest;
-import com.project.Journey.board.entity.Post;
-import jakarta.persistence.criteria.Predicate;
-import org.springframework.data.jpa.domain.Specification;
-import org.springframework.util.StringUtils;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class PostSpecification {
-
-    public static Specification<Post> search(PostSearchRequest request) {
-        return (root, query, builder) -> {
-            List<Predicate> predicates = new ArrayList<>();
-
-            if (request.getTitle() != null && !request.getTitle().isEmpty()) {
-                predicates.add(builder.like(root.get("title"), "%" + request.getTitle() + "%"));
-            }
-
-            if (request.getNickname() != null && !request.getNickname().isEmpty()) {
-                predicates.add(builder.equal(root.get("user_id"), request.getNickname()));
-            }
-
-            if (request.getCountry() != null && !request.getCountry().isEmpty()) {
-                predicates.add(builder.equal(root.get("country"), request.getCountry()));
-            }
-
-            if (request.getPostId() != null) {
-                predicates.add(builder.equal(root.get("postId"), request.getPostId()));
-            }
-
-            if (request.getStartDate() != null) {
-                predicates.add(builder.greaterThanOrEqualTo(root.get("startDate"), request.getStartDate()));
-            }
-
-            if (request.getEndDate() != null) {
-                predicates.add(builder.lessThanOrEqualTo(root.get("endDate"), request.getEndDate()));
-            }
-
-            return builder.and(predicates.toArray(new Predicate[0]));
-        };
-    }
-}
