--- src/main/java/com/project/Journey/board/repository/PostImageRepository.java
@@ -1,15 +1,0 @@
-package com.project.Journey.board.repository;
-
-
-import com.project.Journey.board.entity.Post;
-import com.project.Journey.board.entity.PostImage;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface PostImageRepository extends JpaRepository<PostImage, Long> {
-    List<PostImage> findByPost(Post post);
-
-}
