+++ src/main/java/com/project/Journey/community/repository/CommunityRepository.java
@@ -1,9 +1,6 @@
 package com.project.Journey.community.repository;
 
-import com.project.Journey.board.entity.Post;
-import com.project.Journey.community.dto.SearchDTO;
 import com.project.Journey.community.entity.Community;
-import io.lettuce.core.dynamic.annotation.Param;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -11,10 +8,8 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface CommunityRepository extends JpaRepository<Community, Long>, JpaSpecificationExecutor<Community> {
