+++ src/main/java/com/project/Journey/companion/comment/repository/CommentRepository.java
@@ -1,0 +1,13 @@
+package com.project.Journey.companion.comment.repository;
+
+
+import com.project.Journey.companion.comment.entity.Comment;
+import com.project.Journey.companion.entity.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+    List<Comment> findByPostAndParentCommentIsNullAndIsActiveTrueOrderByCreatedAtAsc(Post post);
+    List<Comment> findByParentCommentAndIsActiveTrueOrderByCreatedAtAsc(Comment parent);
+}
