+++ src/main/java/com/project/Journey/companion/entity/PostImage.java
@@ -1,0 +1,28 @@
+package com.project.Journey.companion.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Table(name = "post_images")
+public class PostImage {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long postImageId;
+
+    //이미지 url
+    @Column(nullable = false)
+    private String postImageUrl;
+
+    //post와의 관계
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id", nullable = false)
+    private Post post;
+
+}
