+++ src/main/java/com/project/Journey/companion/exception/PostException.java
@@ -1,0 +1,16 @@
+package com.project.Journey.companion.exception;
+
+import org.springframework.http.HttpStatus;
+
+public class PostException extends RuntimeException{
+    private final HttpStatus status;
+
+    public PostException(String message, HttpStatus status){
+        super(message);
+        this.status=status;
+    }
+
+    public HttpStatus getStatus(){
+        return status;
+    }
+}
